+++ lib/components/touch_scale_context.dart
@@ -1,9 +1,9 @@
-import 'package:flutter/animation.dart';
-import 'package:flutter/foundation.dart';
+import 'package:flutter/widgets.dart';
 import 'package:flutter_touch_scale/flutter_touch_scale.dart';
 
 @protected
 mixin TouchScaleContext {
+  BuildContext get context;
   TickerProvider get vsync;
   Duration get duration;
   Curve get curve;
@@ -12,4 +12,5 @@ mixin TouchScaleContext {
   Duration? get previewDuration;
   TouchScaleCallPhase get callPhase;
   TouchScaleBehavior get behavior;
+  TouchScaleRejectBehavior get rejectBehavior;
 }